Removed CreativeTab functionality
CreativeTabs only work on the client side, but I was accessing them from the server side because of the way I make my registry. Either I make a server variant of the mod or I disable the functionality until I figure it out properly, and I went with the latter.